requirements.txt
# requirements.txt
Flask>=2.3 
requests>=2.25.0
python-dotenv>=0.19.0
google-generativeai>=0.5.0
PyYAML>=5.0
simple-salesforce>=1.11.0
cryptography>=3.4.0 
PyJWT>=2.0.0
